pyproject.toml
[project]
name = "simpleclaw"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Vincent Liu", email = "zhuoxinliu01@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "croniter>=6.0.0",
    "curl-cffi>=0.14.0",
    "lxml>=6.0.2",
    "openai>=2.28.0",
    "python-dotenv>=1.2.2",
    "python-telegram-bot>=22.7",
    "requests>=2.32.5",
    "wecom-aibot-sdk-python>=0.1.6",
]

